MicroscopeControl

Documentation for MicroscopeControl.

MicroscopeControl.MicroscopeControl

MicroscopeControl.AbstractInstrument

MicroscopeControl.AbstractSystem

MicroscopeControl.AbstractSystemState

MicroscopeControl.export_state

MicroscopeControl.export_state

MicroscopeControl.get_state

MicroscopeControl.initialize

MicroscopeControl.initialize

MicroscopeControl.save_attributes_and_data

MicroscopeControl.save_group_recursive

MicroscopeControl.save_h5

MicroscopeControl.set_state

MicroscopeControl.shutdown

MicroscopeControl.shutdown

MicroscopeControl.MicroscopeControl — Module

MicroscopeControl is the main module for the MicroscopeControl package. It exports all abstract types and methods for instruments, as well as the implementations for all hardware interfaces. It also provides methods for saving data to HDF5 files, and the GUI methods for user interfaces.

Uses Reexport.jl to cleanly propagate exports from submodules.

MicroscopeControl.AbstractInstrument — Type

AbstractInstrument is an abstract type that all instruments should inherit from.

MicroscopeControl.AbstractSystem — Type

AbstractSystem

Abstract type for a microscope system (composite of instruments). Systems compose multiple instruments and manage their collective state.

MicroscopeControl.AbstractSystemState — Type

AbstractSystemState

Abstract type for system state. Concrete implementations should contain the configuration parameters for a specific system.

MicroscopeControl.export_state — Method

export_state(instrument::AbstractInstrument)

Export the state of the instrument as a tuple of attributes, data, and children.

Arguments

instrument::AbstractInstrument: The instrument to export the state of.

Returns

A tuple of dictionaries containing the attributes, data, and children of the instrument.

MicroscopeControl.export_state — Method

export_state(sys::AbstractSystem)

Export the state of the system as a tuple of attributes, data, and children.

Arguments

sys::AbstractSystem: The system to export state from.

Returns

A tuple of (attributes::Dict, data, children::Dict).

MicroscopeControl.get_state — Method

get_state(sys::AbstractSystem)

Get the current state of the system.

Arguments

sys::AbstractSystem: The system to get state from.

Returns

An AbstractSystemState representing the current configuration.

MicroscopeControl.initialize — Method

initialize(instrument::AbstractInstrument)

Initialize the instrument.

Arguments

instrument::AbstractInstrument: The instrument to initialize.

MicroscopeControl.initialize — Method

initialize(sys::AbstractSystem)

Initialize all instruments in the system.

Arguments

sys::AbstractSystem: The system to initialize.

MicroscopeControl.save_attributes_and_data — Method

save_attributes_and_data(filename::String, group::String, attributes::Dict, data::Any, children::Dict{String,Any})

Saves hierarchical data to an HDF5 file.

Arguments

filename::String: Path to the HDF5 file

group::String: Name of the HDF5 group

attributes::Dict: Dictionary of attributes to save

data::Any: Data to save in the group

children::Dict{String,Any}: Dictionary of child groups

MicroscopeControl.save_group_recursive — Method

save_group_recursive(h5parent, group_name::String, attributes::Dict, data, children)

Recursively saves hierarchical data to an HDF5 file.

Arguments

h5parent: Parent HDF5 group

group_name::String: Name of the HDF5 group

attributes::Dict: Dictionary of attributes to save

data::Any: Data to save in the group

children::Dict{String,Any}: Dictionary of child groups

MicroscopeControl.save_h5 — Method

save_h5(filename::String, state_data)

Save the state data to an HDF5 file.

Arguments

filename::String: Path to the HDF5 file

state_data: Tuple of dictionaries containing the attributes, data, and children of the instrument.

MicroscopeControl.set_state — Method

set_state(sys::AbstractSystem, state::AbstractSystemState)

Set the state of the system, updating all relevant instruments.

Arguments

sys::AbstractSystem: The system to configure.

state::AbstractSystemState: The desired state.

MicroscopeControl.shutdown — Method

shutdown(instrument::AbstractInstrument)

Shuts down the instrument.

Arguments

instrument::AbstractInstrument: The instrument to shut down.

MicroscopeControl.shutdown — Method

shutdown(sys::AbstractSystem)

Shutdown all instruments in the system.

Arguments

sys::AbstractSystem: The system to shut down.